requirements.txt
Django==3.1.3
djangorestframework==3.12.2
Markdown==3.3.3
Pillow==8.0.1
python-dateutil==2.8.1
pytz==2020.4
pre-commit
mysqlclient
